refactor(direct-ptycho): extract DirectPtychographyBase, and fix three bugs
Splits `DirectPtychography` so a second, real-space reconstruction class can share its
machinery, and fixes the bugs that surfaced while doing so.

Refactor (behaviour-preserving):
- new `DirectPtychographyBase` holds what does not depend on how the deconvolution is
  performed: the geometry/sampling properties, `HyperparameterState`, bright-field mask
  indexing, `visualize`, and the optuna / grid-search drivers. Subclasses supply
  `reconstruct`, `variance_loss` and `corrected_bf`; the contract is documented on the
  class and declared as annotations.
- `build_vbf_stack_from_dataset4d` and friends move to `direct_ptycho_utils`, so any
  class consuming a virtual bright-field stack runs the same origin-correction,
  masking and normalization pipeline.
- verified by running both trees (`git worktree` at the parent commit) over 5 kernels x
  2 upsampling factors plus variance losses, half-sets, lateral shifts, grid search and
  the least-squares fit: all 35 outputs bit-identical.

Fixes:
- `visualize`'s scalebar ignored `upsampling_factor`, so an upsampled object reported
  double its true field of view. `reconstruct` now records `_last_upsampling_factor` and
  `_obj_sampling` divides by it.
- `from_virtual_bfs` defaulted `semiangle_cutoff` to None, which always raised a
  `TypeError` in the setter. It is now required, with a real error message.
- `gpts`/`scan_gpts` are plain tuples rather than `torch.Size`, so they round-trip
  through `AutoSerialize` as the type they started as.

Because save/load never worked, no stored file can name the old module path, so the
backwards-compatibility re-exports were dropped.

Also adds `scatter_add_splat`: a batched, device-agnostic bilinear/nearest scatter-add,
the torch counterpart of `imaging_utils.bilinear_kde`'s accumulation stage. It carries a
leading batch axis, offers a drop (`"pad"`) as well as a wrap boundary via clamped
indices with zeroed weights (no host sync), and accumulates the sum of squares needed for
a variance estimate.
